Added internal and cmd folder. Moved handler codes to api folder. Changed handler names with api.
