+++ internal/api/create_api.go
@@ -1,8 +1,8 @@
-package handler
+package api
 
 import (
 	"encoding/json"
-	"github.com/batuhankucukali/istekbin/config"
+	"github.com/batuhankucukali/istekbin/internal/config"
 	"github.com/go-redis/redis/v7"
 	"github.com/google/uuid"
 	"github.com/labstack/echo/v4"
