+++ internal/api/create_api_test.go
@@ -1,7 +1,7 @@
-package handler
+package api
 
 import (
-	"github.com/batuhankucukali/istekbin/config"
+	"github.com/batuhankucukali/istekbin/internal/config"
 	"github.com/labstack/echo/v4"
 	"github.com/stretchr/testify/assert"
 	"net/http"
