+++ internal/api/create_list_api_test.go
@@ -1,4 +1,4 @@
-package handler
+package api
 
 import (
 	"encoding/json"
