+++ internal/api/home_api_test.go
@@ -1,4 +1,4 @@
-package handler
+package api
 
 import (
 	"github.com/labstack/echo/v4"
