+++ internal/api/request_api.go
@@ -1,9 +1,9 @@
-package handler
+package api
 
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/batuhankucukali/istekbin/config"
+	"github.com/batuhankucukali/istekbin/internal/config"
 	"github.com/go-redis/redis/v7"
 	"github.com/google/uuid"
 	"github.com/labstack/echo/v4"
