+++ middleware/limiter.go
@@ -1,7 +1,7 @@
 package middleware
 
 import (
-	"github.com/batuhankucukali/istekbin/config"
+	"github.com/batuhankucukali/istekbin/internal/config"
 	"github.com/labstack/echo/v4"
 	_ "github.com/labstack/echo/v4"
 	"github.com/ulule/limiter/v3"
